// components/user/user-validation-schema.ts
import * as yup from 'yup';

export const userValidationSchema = yup.object().shape({
  firstName: yup.string().required('form:error-last-name-required'),
  lastName: yup.string().required('form:error-first-name-required'),
  email: yup
    .string()
    .email('form:error-email-format')
    .required('form:error-email-required'),
  role: yup.object().required('form:error-role-required')
});